+++ Primitives/CustomAnswer.cs
@@ -1,6 +1,6 @@
 namespace Domain.Primitives;
 
-public enum CustomAnswer: byte
+public enum CustomAnswer: int
 {
     No = 0,
     Yes = 1
